pyproject.toml
[project]
name = "h2s-google-genai-apac-cohort1-final"
version = "0.1.0"
description = "ADK multi-agent sidekick (tasks, schedule, notes) with Flask UI and AlloyDB"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.3.2",
    "flask>=3.0",
    "google-adk>=1.28.1",
    "google-api-python-client>=2.136.0",
    "google-cloud-alloydb-connector[pg8000]>=1.9",
    "google-genai>=1.0",
    "httpx>=0.27",
    "pg8000>=1.31",
    "psycopg[binary]>=3.2",
    "python-dotenv>=1.2",
    "ruff>=0.15.9",
    "sqlalchemy>=2.0",
]
